refactor: split out permalink and full table types
